refactor: Move template logic into individual files
