+++ crates/sherwood/src/lib.rs
@@ -15,7 +15,7 @@ pub use generator::{
 };
 pub use plugins::{ContentParser, ParsedContent, PluginRegistry};
 pub use presentation::styles::StyleManager;
-pub use presentation::templates::TemplateManager;
 pub use server::{run_dev_server, run_dev_server_with_plugins};
 pub use sherwood::Sherwood;
+pub use templates::TemplateManager;
 pub use templates::{partials, partials::ContentItem};
