+++ crates/sherwood/src/presentation/mod.rs
@@ -1,4 +1,3 @@
 pub mod css_processing;
 pub mod pages;
 pub mod styles;
-pub mod templates;
